# Conflicts:
#	hoa-microservice/build.gradle

+++ authentication-and-member-microservice/src/main/java/nl/tudelft/sem/template/authmember/models/HoaModel.java
@@ -2,13 +2,13 @@ package nl.tudelft.sem.template.authmember.models;
 
 public abstract class HoaModel {
     private transient String memberId;
-    private transient int hoaId;
+    private transient long hoaId;
 
     public String getMemberId() {
         return memberId;
     }
 
-    public  int getHoaId() {
+    public long getHoaId() {
         return hoaId;
     }
 }
